Reorganisation + Load harfbuzz if available
